Remove SnapshotCaches
It never worked perfectly, we had few percent of tests failed. Yes, it boosted speed of tests 2 times in some cases, but there is a better option - do not use XCTest for finding and checking elements (there is Real View Hierarchy feature, it doesn't work perfectly either, but it is just not finished yet).