Refactor working with UI
* Public interfaces:
`assert.hasText` was changed to `assertHasText`, as well as any other assert (e.g.: `assert.isDisplayed` changed to `assertIsDisplayed`). This will reduce amount of typos, where `assert` was missed (104 times in some app that uses Mixbox).

`visibleText` changed to `text` due to the fact that there seems to be no need to reserve `text` for something else.

New `text()`, `accessibilityLabel()` and generic `value()` methods.

* Extensibility:
It is easier to have custom element types, to add custom checks and actions. Custom checks and actions can access almost anything implemented in Mixbox that will work for both BlackBox and GreyBox tests without effort (and there is an option to inject your own dependencies and manually support BlackBox and GreyBox simultaneously, or just use only one type of testing).

* Actions:
AlmightyElementActions was removed, actions are now split to different classes and they are not dependent on XCUI/GreyBox. Why we had AlmightyElementActions in the first place? Some time ago we had XCUI+EarlGrey implementations and those instruments were very different. Now we want to share most of the code between BlackBox and GreyBox tests, so our internal interfaces can be more clean.

* Checks:
Same thing happened with checks. Now they are very generic and are represented by classes. "Is Not Displayed" check is now simply uses "Is Displayed" check and this allowed to find some false-positive cases in tests of a real application.

* Interactions (Actions + Checks):
The only method of PAge Object Element necessary to start interaction is now `performInteraction`. Checks and actions are now implementations of same protocol (ElementInteraction).

* Other

Tests, refactorings, and other things that exceeds limit on commit message. All those comments will be lost in time, like tears in rain.