Add test for setting notification permissions, improve docs
I've tried to write test, it didn't work, I tried to remember how to integrate MixboxFakeSettingsAppMain into a project... spent whole day, updated docs to make the process easier for other people, add error detection and logs for MixboxFakeSettingsAppMain. For example, I've used private api to detect errors in other private api (completion of setSectionInfo is not called if entitlements are not set in project settings).