Remove "isHidden should never be called for a fake cell" assertion
It fails tests on iOS 13. I have no idea why it was added. When this assertion appeared in iOS 13 I reviewed stacktrace and didn't find anything suspicious. It seems that AX-thing that generates AX hierarchy just gets this property of fake cell, because it was added in an UIAccessibilityContainer (a collection view in that case) and it is expected. Let's just remove the property and see what happens.