Replace some compile time DI with runtime DI based on Dip
This will help to reduce amount of "code smells" in code, remove code duplication (every project that uses Mixbox had to contain a lot of DI code that could be shared for every project), this simplifies development of Mixbox and client projects.

This is not a complete refactoring, just a first step. There is still a lot of mess with DI in Mixbox.

An abstraction for DI was added that hides Dip implementation. It is compatible with pure Dip DI and can be connected to DI in existing project that uses Dip.