Rename `reopen-fast` to `r`
I was calling `reopen` every time, so I got my project cleaned every time. I asked myself, why can't I remember or type `reopen-fast`. And I think that this is because the name is not short. I renamed it to `r`, because it is very fast to type (`make r`) and this is (almost) the only thing I use.