Always require element to be visible 100% when scrolling
The previous way of requiring only `minimalPercentageOfVisibleArea` resulted in bug, when center of the tapped view (coordinate for tap) was overlapped by another view even if tapped view was visible by more than 50%. This was happening when overlapping view's width was less than scroll view width (for case with vertical scrolling). So it was overlapping center of target view, but target view was visible by more than 50%.

The solution was to always require 100% visibility before finishing scrolling. Currently if view is centered in scroll view, the scrolling ends. And the 100% requirement is only set for scrolling, so if view is actually visible by 95% for example, the interaction (action or check) will continue as usual (as before this fix).