Restore SurrogateCAAnimationDelegate
It is not practically possible to use simple wrapper of CAAnimationDelegate. There are a lot of downcasts and forcecasts in UIKit. It is required that we do not change instances of CAAnimationDelegate.

For example, `_transferAnimationToTrackingAnimator` contains this code (simplified):

```
rbx = [[r14 delegate] retain];
r15 = r13->_nextState;

COND = rbx == r15;
if (!COND) {
    ...assertion failure...
```

Here in this UIKit function there is an assumption that `_nextState` of type `UIViewAnimationState` points to `delegate` of type `CAAnimationDelegate`. This is kind of a forcecast and a violation of multiple OOP patterns.

I tried to make a workaround, but there are more places, for example, this:

```
int ____UIViewAnimationDelegateAllowsUserInteraction_block_invoke(int arg0, int arg1, int arg2, int arg3) {
    var_38 = arg3;
    r12 = arg0;
    rbx = [arg1 retain];
    rax = [*(r12 + 0x20) layer];
    rax = [rax retain];
    r13 = rax;
    var_30 = rbx;
    rax = [rax animationForKey:rbx];
    rax = [rax retain];
    r14 = [[rax delegate] retain];
    [rax release];
    [r13 release];
    if ((objc_opt_respondsToSelector(r14, @selector(_allowsUserInteraction)) & 0x1) != 0x0) {
            *(int8_t *)(*(*(r12 + 0x28) + 0x8) + 0x18) = [r14 _allowsUserInteraction];
    }
    r15 = var_30;
    if ((*(int8_t *)(*(*(r12 + 0x28) + 0x8) + 0x18) == 0x0) && (((objc_opt_respondsToSelector(r14, @selector(_allowsUserInteractionToCutOffEndOfAnimation)) & 0x1) != 0x0) && ([r14 _allowsUserInteractionToCutOffEndOfAnimation] != 0x0))) {
            *(int8_t *)(*(*(r12 + 0x28) + 0x8) + 0x18) = __UIViewLayerAnimationCanBeConsideredFinished(*(r12 + 0x20), r15);
    }
    *(int8_t *)var_38 = *(int8_t *)(*(*(r12 + 0x28) + 0x8) + 0x18) ^ 0x1;
    [r14 release];
    rax = [r15 release];
    return rax;
}
```

There is a downcast to `UIViewAnimationState` (actually, not to a class, but to a selector, kind of a downcast). And this code doesn't handle the case properly if `CAAnimationDelegate` is not `UIViewAnimationState`, it prevents interactivePopGesture from working, multiple gestures causes UI to stuck.

The solution is to bring back `SurrogateCAAnimationDelegate` (as in EarlGrey). It was causing crashes recently, because of incorrect bringing to Obj-C and incorrect translation to Swift, but it was easy to fix actually.