Add class that can be used in tests to generate stubs with minimal boilerplate.
`GeneratorFacade` is created as "easy to use class for generating stubes". See tests. Here's one example of usage:

```
book = generator.generate {
    $0.id = 1
    $0.title = "Hamlet"
    $0.author.stub {
        $0.id = 2
        $0.name = "William Shakespeare"
    }
}
```

The ideas behind the implementation were:
- Minimize boilerplate as much as possible
- Make everything as automatic as possible
- Make everything as configurable as possible

Example: there is no need to stub a class `Author` to fill required instance field `author` in class `Book`. Same for arrays, dictionaries, optionals.

Additional requirerd changes:
- Make DI more configurable, combinable, etc
- Refactor DI
- Add classes that convert Swift errors (exceptions) to test failures, this imporves readability of the code and minimizes boilerplate.

This is still a prototype and is not ready for real tests.