Refactor visibility check
The idea was to improve performance by performing less actions. In the future commits there will be an algorithm that check less pixels of image, in this commit I prepared everything and also tried to get rid of intermediate step of taking screenshot data (it did not work). Currently there were screenshotter and it rendered data to image, then image was rendered to XRGB. I tried to make it run in one pass but gave up and maybe this is not possible. Maybe the better idea will be doing everything on GPU without switching to CPU (without transferring and converting data), maybe it is possible and it will be better.