Replace automatic disabling of visibility check optimization with manual disabling
`with(optimizedVisibilityCheck: false)` modifier was added, default value is true. This is due to the fact that otherwise optimization will be really disabled, because scroller always wants to make view hundred percent visible and automatic disabling disabled optimization for such requirements. This would ruin performance if was kept as is.