Make generator interfaces module lightweight
This change is required to minimize boilerplate in tests. Linking `MixboxGenerators` to a module with models allows to not write initializers for models (e.g. if they are inherited from `Unboxable` and have no initializers with plain fields), and every initializer requires at least 2*N lines of code where N is number of fields in a model. In case of writing an extension for your model and conforming it to `InitializeableByFields`, you have to write 3*N lines of code. With initializing model directly from `Fields` you have to write 1*N lines. Also when `MixboxGenerators` is linked directly to your framework with models it will be much easier for you to write ccode, because you will not have to create a separate module with extensions, you will not even have to create a separate file. Also extensions slow down compilation speed and auto-completion.

Before this commit `MixboxGenerators` couldn't be linked to a module (practically), because it was dependent on `Dip` and `Dip` doesn't have conditional compilation clauses to strip testing code from release build. Also `MixboxGenerators` was dependent on `MixboxTestsFoundation`, which was dependent on `XCTest` and other modules. Now it is very lightweight and has no code that will not be stripped from release build.

Changes:
- `MixboxGenerator` was split into `MixboxGenerators` and `MixboxStubbing`
- `MixboxGenerators` contains everything that can be linked to a module with models, to implement generation of models from stubs, the only dependency is `MixboxDi`, which contains only 8 very small files with 84 lines of code in total.
- `MixboxDi` was split into `MixboxDi` (very lightweight) and `MixboxDipDi` (with `Dip` and `MixboxFoundation` as its only dependencies)