Improve customization of classes for tests ("Testability")
Now every property of element, that is visible from tests, can be overriden. Before that only some properies could be overridden (text, hasKeyboardFocuse and some others).

A protocol `TestabilityElement` was added. It can be used in both Obj-C and Swift. Note that working with custom values requires Swift, because it is made using `Codable`.

The reason for implementing interfaces for Objective-C is not only to make them available in Objective-C, but also because otherwise it was impossible to make overrideable functions in extensions in Swift. It was only possible when using Objective-C categories.

This change will allow to mix AX hierarchy with UIKit hierarchy. For example, it will enable testing web views or custom rendered things like maps.