Support gray box testing non-view elements defined via accessibilityElements in app
Initially I wanted to combine AX and UIKit hierarchies by returning `accessibilityElements` along with `subviews` in `mb_testability_children`. But then I realized again how unpredictable AX hierarchy is and how it differs from UIKit hierarchy. After implementing mix of AX and UIKit hierarchies I got duplicated elements, mixed in a very bizarre way. So I decided to give up on AX hierarchy completely. If we want to implement testing of `WKWebView`, we will implement this feature separately. For example, add extensions to `WKWebView`. And `WKWebView` doesn't even provide its accessibility elements, because rendering is happening in a separate process (same for keyboard), so we either will spend a lot of time disassembling and reverse engineering AXRuntime and other frameworks or give up on this feature.