Add MixboxReflection
It will be used to print automatically generated values in MixboxStubbing.

I was thinking that models for reflection can be mutable and immutable (example: Mirror is mutable), I even created two sets of models, however, I deleted the whole set of mutable models, because they are not used and not tested (the code is only on my local machine). In future they can be added back if they will be needed. The current naming has word "immutable", I think it is fine.