Fix calling default implementation
There was a bug (not like something was working wrong, but rather a bug as something that doesn't match initial idea), second parameter of `defaultImplementationClosure` wasn't used. Nothing was wrong, because currently it is equal to arguments of the called function, however, it can be changed in future, so it's better to use tupled arguments instead of using function's arguments directly.