Get rid of some dependencies on Apple's AX
AX here means some kind of accessibility extensions ("AX" is used a lot in private APIs) that are only available in Apple's UI tests (if we aren't using hacks like in `AccessibilityOnSimulatorInitializer` and they don't fail, which is what actually happens sometimes).

Not everything is done - `accessibilityLabel` and `accessibilityValue` still don't provide same thing in unit tests as in UI tests. This will be addressed later.