Add tests.
ImageHashCalculatorSnapshotsComparatorTests was added with multiple tests.

Another test was added to `ImageHashCalculatorTests` for new behavior (empty images are now handled properly - hash can't be generated for image with no data, so `imageHash` throws an exception)

One of the tests is disabled: `disabled_test___compare___tells_images_are_similar___if_only_difference_is_transparency_of_pixels` This is a test for a feature that will be implemented later.