Fix extreme flakiness of setting notification permissions in Mixbox
The issue was when completion handler of `BBSectionInfoSettings`'s `-setSectionInfo:forSectionID:withCompletion:` was called, there was EXC_BAD_ACCESS.

It was probably happening because of incorrect signature of the block (I am mostly sure).

One of the solutions was to get rid of completion block completely, to use `-setSectionInfo:forSectionID:`.

However, disassembled code of `BulletinBoard.framework` showed that it calls `-setSectionInfo:forSectionID:withCompletion:` with completion `nil`, meaning that can not be synchronized.

I removed all arguments from the block and it worked, for iOS 15 and iOS 12. Disassembled code for iOS 12 and iOS 15 also looks very similar. Tests worked with no problem, the state of norifications permissions (allowed/denied) can be reliably switched based on several runs.

Additional changes:
- Use Objective-C API for interacting with `BBSectionInfoSettings` and `BBSectionInfo`.
- `-isSectionInfoApplied:actualSectionInfo:` is used to check if section info is applied. `BBSettingsGateway`'s `-sectionInfoForSectionID:` method actually goes somewhere via XPC synchronously and it seems that it gets actual state.
- Code for iOS 9 was removed (it is no longer supported).