Fix tapping keyboard on iOS 16
On iOS 16 keyboard done button is no longer present in UIKit hierarchy, view hierarchy now is very limited (also it is no longer shown on screenshots)

The solution was to add additional hierarchy to `ViewHierarchyProviderImpl`, comprised of keyboard and keyboard keys (accessibility elements were used for that)

New keyboard hierarchy was added to the top level of `ViewHierarchy`.

Most of the fields of `ViewHierarchyElement` are calculated ordinarily, using accessibility.

Frames of keys (relative to container) are calculated based on keyboard layout frame (note that keyboard layout is not full keyboard, it misses UI of predictive input)

All keys contain custom value name `MixboxBuiltinCustomValue.UIKBKey.name`. Example values:
- `Latin-Capital-Letter-H`
- `Space-Key`
- `Return-Key`

It was added, because some of existing `accessibilityLabel`'s are locale-dependent, `accessibilityIdentifier` of return key is different for different `UIReturnKeyType`s.

Changes:
- `CanTapKeyboardTests` was added
- Fix `NonViewVisibilityCheckerImpl`, not every element has correct `mb_testability_frame()`
- Improve error description in `ScrollingHintIpcMethodHandler`
- `TestabilityElementType` enum was fixed, first case `TestabilityElementType_Other` should start at 1 instead of 0
- `mb_testability_elementType()` is now calculated  using `_accessibilityAutomationType` for `NSObject`
- `HasNoSuperviewMatcher` was added (in order to match new keyboard element, which has no superview, more accurately)
- Limit usage of `KeyboardPrivateApi.sharedInstance()`, inject it wherever is possible
- Add shorthand operators to `Coordinates+Operators.swift`