Fix keyboard being invisible for InAppScreenshotTaker on iOS 16
This caused Mixbox to "see through" the keyboard, because it wasn't present on screenshot. So Mixbox would tap views even if they were below keyboard, as if they were fully visible.

Note that there are still issues with visibility check. The keyboard is kind of transparent (blurs content behind it) and actually you can see through it, even keyboard visually appears above elements that are behind it.