Replace all magic numbers of iOS versions with values from MixboxIosVersions
This is done to make it easier to remove unsupported code. If we delete a supported version, we can delete all related code.
`MixboxIosVersions` has subtypes: `Supported`. `Outdated`, `Future`.

Code related to `Supported` versions should be present.

Code related to `Outdated` may be removed.

Code related to `Future` is used to check for an iOS version greater than latest supported version.

If version is in neither of those classes, code should definitely be removed.